Time: 12 ms (92.31%), Space: 91.8 MB (96.15%) - LeetHub

+++ 3524-find-x-value-of-array-i/3524-find-x-value-of-array-i.java
@@ -7,7 +7,7 @@ class Solution {
             long[] d = new long[k];
             d[c]++; a[c]++;
             for (int j = 0; j < k; j++) {
-                int n=(int) (((long) j*c)%k);
+                int n=(j*c)%k;
                 d[n]+=b[j];
                 a[n]+=b[j];
             }
